% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_bangladesh}
\alias{dengue_bangladesh}
\title{Dengue Dataset of Bangladesh: Unraveling Dengue}
\format{
A dataframe object with 1000 rows and 19 variables:
\describe{
  \item{gender}{Gender of the survey respondent}
  \item{age}{Age of the survey respondent}
  \item{ns1}{Non-Structural Protein 1  test's data positive as 1 negative as 0}
  \item{ig_g}{IgG data positive as 1 negative as 0}
  \item{ig_m}{IgM data positive as 1 negative as 0}
  \item{area}{The area within the Dhaka region where the respondent resides.}
  \item{area_type}{Classification of the area as "Developed" or "Undeveloped}
  \item{house_type}{The type of housing in the area, e.g., "Building," "Tin-Shed," or "Others."}
  \item{district}{The specific district within Dhaka}
  \item{fever_duration}{ Acute Dengue febrile phase typically lasts 2–7 days}
  \item{body_temperature}{Biphasic, high fever is the primary clinical entry criterion}
  \item{platelet_count}{Thrombocytopenia (platelets < 150,000) is a critical indicator.}
  \item{wbc_count}{Leukopenia (low WBC count) is highly characteristic of acute Dengue}
  \item{joint_pain}{Arthralgia is so severe it gave Dengue the name "breakbone fever"}
  \item{headache}{Frontal headache is common in acute viral syndromes.}
  \item{retro_orbital_pain}{Extremely specific indicator for Dengue vs. other febrile illnesses}
  \item{myalgia}{Generalized muscle pain is characteristic.}
  \item{rash}{Transient maculopapular rash occurs during convalescence.}
  \item{outcome}{Outcome variable indicating the presence (1) or absence (0) of Dengue cases reported}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/kawsarahmad/dengue-dataset-bangladesh, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_bangladesh
}
\description{
This dataset presents real-world data collected through surveys conducted in the Dhaka region of Bangladesh. It focuses on understanding the prevalence and characteristics of the Dengue fever phenomenon, a significant public health concern in the area. The dataset is updated monthly to reflect the evolving nature of the Dengue outbreak.
}
\details{
The dengue_bangladesh dataset has also been published in the denguedataset package in Python. You can find the Zenodo page https://zenodo.org/records/21499199, GitHub repository https://github.com/fdzul/denguedatasets-py, GitHub Pages https://fdzul.github.io/denguedatasets-py/, and PyPI index at the following links https://pypi.org/project/denguedatasets/.
}
\keyword{datasets}
