% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_brazil}
\alias{dengue_brazil}
\title{Brazil Dengue Dataset 2000-2019}
\format{
A dataframe object with 26 rows and 4 variables:
\describe{
  \item{micro_code}{Unique ID number given by IBGE to each microregion (5 digits)}
  \item{micro_name}{Name of each microregion}
  \item{micro_name_ibge}{Name of each microregion from IBGE sources}
  \item{meso_code}{Unique ID number given by IBGE to each mesoregion: 4 digits}
  \item{meso_name}{Name of each mesoregion}
  \item{state_code}{Unique ID number given by IBGE to each state}
  \item{state_name}{Name of state}
  \item{region_code}{Region code}
  \item{region_name}{Region name}
  \item{biome_code}{Biome code}
  \item{biome_name}{Biome name}
  \item{ecozone_code}{Ecozone code: same as biome apart from Atlantic Rainforest, biome code 6, which has been separated}
  \item{ecozone_name}{Ecozone name: same as biome apart from Atlantic Rainforest, which has been separated into NE..}
  \item{main_climate}{Most prevalent climate regime in the microregion. Based on Koppen Geiger climate regimes}
  \item{month}{Calendar month index, 1 = January, to, 12 = December}
  \item{year}{Year: 2000-2019}
  \item{time}{Time index starting at 1 for January 2000}
  \item{dengue_cases}{Number of notified dengue cases registered in the notifiable diseases system in Brazil by SINAN}
  \item{population}{Estimated population, based on projections calculated using the 2000 and 2010 censuses, and counts}
  \item{pop_density}{Population density: number of people per km2}
  \item{tmax}{Monthly average daily maximum temperature; gridded values at a 0.5° resolution averaged across eac}
  \item{tmin}{Monthly average daily minimum temperature; gridded values at a 0.5° resolution averaged across eac}
  \item{pdsi}{Self-calibrated Palmer drought severity index for each microregion}
  \item{urban}{Percentage of inhabitants living in urban areas according to the 2010 census}
  \item{water_network}{Percentage of inhabitants with access to the piped water network according to the 2010 census}
  \item{water_shortage}{Frequency of reported water shortages per microregion between 2000 - 2016}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/raomuhammadsaeedali/brazil-dengue-dataset-2000-2019, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_brazil
}
\description{
The dataset supplied comprises a comprehensive collection of information pertaining to numerous geographical and environmental characteristics across microregions in Brazil from 2000 to 2019. Microregion codes and names, mesoregion codes and names, state codes and names, region codes and names, biome codes and names, ecozone codes and names, climate regimes, months, years, times, dengue cases, population estimates, population density, maximum and minimum temperatures, Palmer's drought severity index, urban population percentages, access to water network percentages, and reported water shortage frequency are all included in the dataset. This information is linked to individual microregions and provides insights into population dynamics, climatic patterns, urbanization trends, water resources, and disease occurrences.
}
\details{
The dengue_brazil dataset has also been published in the denguedataset package in Python
}
\keyword{datasets}
