% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_brazil_labels}
\alias{dengue_brazil_labels}
\title{Brazil Dengue Dataset 2000-2019 labels}
\format{
A dataframe object with 26 rows and 4 variables:
\describe{
  \item{variable}{variable name}
  \item{long_name}{long variable name}
  \item{description}{variable description}
  \item{source}{source}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/raomuhammadsaeedali/brazil-dengue-dataset-2000-2019/data?select=data_desc.csv, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_brazil_labels
}
\description{
Data dictionary/variable-label reference table accompanying the Brazil Dengue Dataset 2000-2019, documenting field names, codes, and descriptions used across microregion, mesoregion, state, region, biome, and ecozone identifiers, as well as climatic and demographic variables in the main dataset.
}
\details{
The dengue_brazil_labels dataset has also been published in the denguedataset package in Python
}
\keyword{datasets}
