% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_colombia}
\alias{dengue_colombia}
\title{Dengue Tabular Data and Satellite Images}
\format{
A dataframe object with 1121 rows and 1017 variables:
\describe{
  \item{municipality_code}{Unique identifier of the municipality}
  \item{municipality}{name of the municipality}
  \item{population}{population:2007-2019}
  \item{cases}{dengue cases: 2007-2019}
  \item{age0_4_percent}{Age0-4 percent }
  \item{age5_14_percent}{Age5-14 percent}
  \item{age15_29_percent}{Age 15 to 29 percent}
  \item{age_30_percent}{Age >30 percent}
  \item{afrocolombian_population_percent}{Percentage of Afro-Colombian population}
  \item{indian_population_percent}{Percentage of Indian population}
  \item{peoplewith_disabilities_percent}{Percentage of people with disabilities}
  \item{peoplewhocannotreadorwrite_percent}{Percentage of people who cannot read or write}
  \item{secondary_higher_education_percent}{Percentage of people who have a secondary/higher education level}
  \item{employedpopulation_percent}{Percentage of employed population}
  \item{unemployedpopulation_percent}{Percentage of unemployed population}
  \item{peopledoinghousework_percent}{Percentage of people who do domestic work}
  \item{retiredpeople_percent}{Percentage of retired people}
  \item{men_percent}{Gender or population percentage of men }
  \item{women_percent}{Gender or population percentage of women}
  \item{householdswithoutwateraccess_percent}{Households without access to water}
  \item{householdswithoutinternetaccess_percent}{Households without internet access}
  \item{buildingstratification1_percent}{Housing stratification between 1}
  \item{buildingstratification2_percent}{Housing stratification between 2}
  \item{buildingstratification3_percent}{Housing stratification between 3}
  \item{buildingstratification4_percent}{Housing stratification between 4}
  \item{buildingstratification5_percent}{Housing stratification between 5}
  \item{buildingstratification6_percent}{Housing stratification between 6}
  \item{numberofhospitalsper_km2}{Number of hospitals per km2}
  \item{numberofhousesper_km2}{Number of houses per km2}
  \item{temperature}{temperature: ene-dec 2007-2018}
  \item{precipitation}{precipitation: ene-dec 2007-2018}
  \item{x2007}{ Year/epiweek: year 2007-2019, week 1 to 52}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/davidrestrepo/dengue, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_colombia
}
\description{
Dataset with sociodemographic, socioeconomic data, dengue cases, temperature and precipitation for all municipalities
}
\details{
The dengue_colombia dataset has also been published in the denguedataset package in Python
}
\keyword{datasets}
