% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_colombia_medellin}
\alias{dengue_colombia_medellin}
\title{Dengue Tabular Data and Satellite Images}
\format{
A dataframe object with 624 rows and 10 variables:
\describe{
  \item{cases_medellin}{dengue cases}
  \item{data}{year-week}
  \item{year}{year}
  \item{year_week}{yearweek:}
  \item{last_day_week}{Last day week}
  \item{month}{month}
  \item{precipitation_medellin}{precipitation}
  \item{precipitation_medellin_rural}{precipitation rural}
  \item{precipitation_medellin_urbano}{precipitation urban}
  \item{temperature}{temperature}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/davidrestrepo/dengue, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_colombia_medellin
}
\description{
Dataset with weekly Dengue cases, temperature and precipitation in Medellin using the epidemiological week.
}
\details{
Dengue Challenge 2 - Make Health LATAM 2022. This challenge will provide a dataset for each municipality in Colombia, composed of weekly dengue cases between 2007 and 2019, along with other social and climatological variables. You will also find a special dataset we have prepared for the municipality of Medellín, with preprocessed climatological variables and a compressed version of satellite images for inclusion in the models. The dengue_colombia_medellin dataset has also been published in the denguedataset package in Python
}
\keyword{datasets}
