% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_india}
\alias{dengue_india}
\title{Dengue cases and deaths in India}
\format{
A dataframe object with 36 rows and 10 variables:
\describe{
  \item{num}{numeric id}
  \item{states}{state of India}
  \item{x2019_cases}{dengue cases of 2019}
  \item{x2019_cases}{deaths cases of 2019}
  \item{x2020_deaths}{dengue cases of 2020}
  \item{x2020_cases}{deaths cases of 2020}
  \item{x2021_deaths}{dengue cases of 2021}
  \item{x2021_cases}{deaths cases of 2021}
  \item{x2022_deaths}{dengue cases of 2022}
  \item{x2022_cases}{deaths cases of 2022}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/jadhavpranav/dengue-cases-in-india, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_india
}
\description{
This data contains dengue cases and deaths happening in each state of India over the years
}
\details{
This data is scrapped from the National Center for Vector Borne Diseases Control. It's a website managed by the Ministry of Health & Family Welfare, Government of India. This data contains dengue cases and deaths happening in each state of India over the years. The dengue_india dataset has also been published in the denguedataset package in Python
}
\keyword{datasets}
