% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_jakarta}
\alias{dengue_jakarta}
\title{Jakarta Dengue Cases and Weather Data from 2021 to 2024.}
\format{
A dataframe object with 2922 rows and 14 variables:
\describe{
  \item{id}{numeric id}
  \item{reg}{Region of Regency or City}
  \item{date}{Date of observation}
  \item{tn}{Minimum temperature in Celsius}
  \item{tx}{Maximum temperature in Celsius}
  \item{tavg}{mean temperature in Celsius}
  \item{rh_avg}{Average relative humidity in percent}
  \item{rr}{Rainfall in mm}
  \item{ss}{Sunshine duration in hours}
  \item{ff_x}{Maximum wind speed in meters by seconds)}
  \item{ddd_x}{Wind direction at maximum wind speed}
  \item{ff_avg}{Average wind speed in meters by second)}
  \item{ddd_car}{Most frequent wind direction}
  \item{cases}{Number of hospitalized patients}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/yutakatarokusumah/jakarta-dengue-weather-2021-2024, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_jakarta
}
\description{
This data contains dengue cases hospitalized of Jakarta
}
\details{
The dengue_jakarta dataset has also been published in the denguedataset package in Python. jupyter notebook related https://www.kaggle.com/code/yutakatarokusumah/jakarta-dengue-risk-prediction
}
\keyword{datasets}
