% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_pakistan}
\alias{dengue_pakistan}
\title{Dengue surveillance data of Pakistan.}
\format{
A dataframe object with 420 rows and 5 variables:
\describe{
  \item{month}{Month name: January through December}
  \item{year}{Year of observation from 2016 to 2020}
  \item{region}{Province or territory name}
  \item{dengue_cases}{Total number of confirmed dengue fever cases reported}
  \item{dengue_deaths}{Total number of deaths attributed to dengue fever}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/aamir28/pakistan-dengue-surveillance-data-2016-2020, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_pakistan
}
\description{
This dataset contains monthly dengue fever surveillance data from Pakistan covering the period from 2016 to 2020.
}
\details{
This dataset contains monthly dengue fever surveillance data from Pakistan covering the period 2016-2020. It provides comprehensive information on dengue cases and deaths across all provinces and territories, making it valuable for epidemiological research, public health policy development, and disease outbreak analysis. Time Period: January 2016 to December 2020 (5 years). Geographic Coverage: All provinces and territories of Pakistan. Frequency: Monthly aggregated data. Total Records: 420 monthly observations. Geographic Regions Included Punjab: Most populous province, Sindh: Second largest province including Karachi, KPK (Khyber Pakhtunkhwa): Northwestern province, Balochistan: Largest province by area, ICT (Islamabad Capital Territory): Federal capital territory, AJK (Azad Jammu & Kashmir): Administered territory, GB (Gilgit-Baltistan): Northern territory.The dengue_pakistan dataset has also been published in the denguedataset package in Python. jupyter notebook related https://www.kaggle.com/code/yutakatarokusumah/jakarta-dengue-risk-prediction
}
\keyword{datasets}
