% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_philippines}
\alias{dengue_philippines}
\title{Philippines Dengue Cases from 2016 to 2020}
\format{
A dataframe object with 1020 rows and 5 variables:
\describe{
  \item{month}{Month of the year in text format}
  \item{year}{Ranges from 2016 to 2020 in numerical format}
  \item{region}{Region in the Philippines}
  \item{dengue_cases}{Number of Monthly Cases per region}
  \item{dengue_deaths}{Number of Monthly Deaths per region due to dengue}
  ...
}
}
\source{
\url{https://www.kaggle.com/datasets/vincentgupo/dengue-cases-in-the-philippines, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_philippines
}
\description{
This dataset contains monthly and regional cases of dengue from 2016 to 2020
}
\details{
Data set contains the recorded number of dengue cases per region of the Philippines from year 2016 to 2020. It can be used to find trends about the disease as well as spatiotemporal analysis that can result into data-driven solution about the trends of the desease for the past 5 years. The dengue_philippines dataset has also been published in the denguedataset package in Python. You can find the Zenodo page https://zenodo.org/records/21499199, GitHub repository https://github.com/fdzul/denguedatasets-py, GitHub Pages https://fdzul.github.io/denguedatasets-py/, and PyPI index at the following links https://pypi.org/project/denguedatasets/.
}
\keyword{datasets}
