% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_sierra_leone}
\alias{dengue_sierra_leone}
\title{The Sierra Leone Freetown DengueAI Dataset from 2015 to 2024}
\format{
A dataframe object with 120 rows and 5 variables:
\describe{
  \item{date}{date}
  \item{temperature_c}{Temperature in celsius}
  \item{humidity_percent}{Humidity}
  \item{precipitation_mm}{Precipitation}
  \item{dengue_cases}{dengue cases}
  ...
}
}
\source{
\url{ - https://www.kaggle.com/datasets/michaalmarkosesay/sierra-leone-freetown-dengueai-dataset-20152024 - https://zenodo.org/records/21499199, - https://fdzul.github.io/denguedatasets-py/ - https://fdzul.github.io/denguedatasets-py/  - https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_sierra_leone
}
\description{
The dataset captures monthly variations in temperature, humidity, and precipitation, as well as confirmed dengue cases in Freetown, Sierra Leone.
}
\details{
The Sierra Leone Freetown DengueAI Dataset from 2015 to 2024 offers an integrated collection of epidemiological and climatic variables, specifically designed for AI-based dengue outbreak prediction and time-series modeling. The dataset captures monthly variations in temperature, humidity, and precipitation, as well as confirmed dengue cases in Freetown, Sierra Leone. Data sources include OpenDengue and NOAA GSOD.The dengue_sierra_leone dataset has also been published in the denguedataset package in Python. You can find the [Zenodo page](https://zenodo.org/records/21499199), [GitHub repository](https://github.com/fdzul/denguedatasets-py), [GitHub Pages](https://fdzul.github.io/denguedatasets-py/), and PyPI index at the following links https://pypi.org/project/denguedatasets/.
}
\author{
Michael Marko Sesay. 2025. Sierra Leone Freetown DengueAI Dataset (2015–2024) [Data set]. Kaggle. https://doi.org/10.34740/KAGGLE/DSV/13257213
}
\keyword{datasets}
