% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengue_bangladesh_mortality}
\alias{dengue_bangladesh_mortality}
\title{Climatic Influence on Dengue Mortality}
\format{
A dataframe object with 150 rows and 15 variables:
\describe{
  \item{division}{Divisions: Chittagong, Rajshahi, Khulna, Sylhet, Barishal, Rangpur}
  \item{year}{year covered: 2000-2024}
  \item{max_temp}{Highest daily temperature influencing mosquito activity and virus transmission}
  \item{min_temp}{Lowest daily temperature affecting mosquito survival and breeding.}
  \item{max_wind_speed}{Measures airflow strength, which impacts mosquito movement and dispersion}
  \item{min_wind_speed}{Measures airflow strength, which impacts mosquito movement and dispersion}
  \item{precipitation_corrected_sum}{Total rainfall contributing to water accumulation for mosquito breeding sites}
  \item{all_sky_surface_total_par}{}
  \item{root_zone_soil_wetness}{}
  \item{surface_soil_wetness}{Moisture level in the soil reflecting wet conditions favorable for breeding habitats.}
  \item{humidity}{Atmospheric moisture level that increases mosquito lifespan and biting frequency}
  \item{earth_skin_temp}{Land surface temperature indicating environmental heat that accelerates mosquito development}
  \item{dengue_national_cases}{Number of reported dengue infections in a given period.}
  \item{dengue_national_deaths}{Total number of dengue-related fatalities}
  \item{dengue_national_cfr_pct}{Case Fatality Rate showing the percentage of infected individuals who died}
  ...
}
#' @source \url{https://www.kaggle.com/datasets/gazijahin/climatic-influence-on-dengue-mortality-2000-2024, https://zenodo.org/records/21499199, https://fdzul.github.io/denguedatasets-py/, https://fdzul.github.io/denguedatasets-py/, https://pypi.org/project/denguedatasets/.}
}
\usage{
dengue_bangladesh_mortality
}
\description{
This dataset explores the spatio-temporal interaction between climate variability and dengue outbreaks across six divisions of Bangladesh from 2000 to 2024. It connects meteorological parameters with epidemiological indicators (cases, deaths, and case-fatality rate) to uncover how changing weather patterns influence dengue mortality
}
\details{
The dengue_bangladesh dataset has also been published in the denguedataset package in Python. You can find the Zenodo page https://zenodo.org/records/21499199, GitHub repository https://github.com/fdzul/denguedatasets-py, GitHub Pages https://fdzul.github.io/denguedatasets-py/, and PyPI index at the following links https://pypi.org/project/denguedatasets/.
}
\keyword{datasets}
